+++ src/main/java/net/detalk/api/image/domain/AttachmentFile.java
@@ -1,4 +1,4 @@
-package net.detalk.api.domain;
+package net.detalk.api.image.domain;
 
 import lombok.Builder;
 import lombok.Getter;
