+++ src/main/java/net/detalk/api/image/repository/AttachmentFileRepository.java
@@ -1,6 +1,6 @@
-package net.detalk.api.repository;
+package net.detalk.api.image.repository;
 
-import net.detalk.api.domain.AttachmentFile;
+import net.detalk.api.image.domain.AttachmentFile;
 
 public interface AttachmentFileRepository {
     AttachmentFile save(AttachmentFile file);
