+++ src/main/java/net/detalk/api/image/service/FileService.java
@@ -1,10 +1,10 @@
-package net.detalk.api.service;
+package net.detalk.api.image.service;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import net.detalk.api.domain.AttachmentFile;
-import net.detalk.api.domain.PreSignedData;
-import net.detalk.api.repository.AttachmentFileRepository;
+import net.detalk.api.image.domain.AttachmentFile;
+import net.detalk.api.image.domain.PreSignedData;
+import net.detalk.api.image.repository.AttachmentFileRepository;
 import net.detalk.api.support.TimeHolder;
 import net.detalk.api.support.UUIDGenerator;
 import net.detalk.api.support.s3.StorageClient;
