+++ src/main/java/net/detalk/api/image/service/ImageService.java
@@ -1,11 +1,11 @@
-package net.detalk.api.service;
+package net.detalk.api.image.service;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import net.detalk.api.domain.AttachmentFile;
-import net.detalk.api.domain.UploadImageData;
-import net.detalk.api.domain.exception.InvalidImageFormatException;
-import net.detalk.api.repository.AttachmentFileRepository;
+import net.detalk.api.image.domain.AttachmentFile;
+import net.detalk.api.image.controller.response.UploadImageData;
+import net.detalk.api.image.domain.exception.InvalidImageFormatException;
+import net.detalk.api.image.repository.AttachmentFileRepository;
 import net.detalk.api.support.TimeHolder;
 import net.detalk.api.support.UUIDGenerator;
 import net.detalk.api.support.image.ImageClient;
