+++ src/main/java/net/detalk/api/post/repository/ProductPostIdempotentRepository.java
@@ -1,0 +1,10 @@
+package net.detalk.api.post.repository;
+
+import java.time.Instant;
+import java.util.UUID;
+
+public interface ProductPostIdempotentRepository {
+
+    boolean insertIdempotentKey(UUID idempotentKey, Instant now);
+
+}
