+++ src/main/java/net/detalk/api/post/service/ProductPostIdempotentService.java
@@ -4,7 +4,7 @@ import java.time.Instant;
 import java.util.UUID;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import net.detalk.api.repository.ProductPostIdempotentRepository;
+import net.detalk.api.post.repository.ProductPostIdempotentRepository;
 import org.springframework.stereotype.Service;
 
 @Slf4j
