+++ src/main/java/net/detalk/api/service/AuthService.java
@@ -3,6 +3,7 @@ package net.detalk.api.service;
 import java.util.UUID;
 import net.detalk.api.controller.v1.response.SessionInfoResponse;
 import net.detalk.api.domain.*;
+import net.detalk.api.image.domain.AttachmentFile;
 import net.detalk.api.member.domain.LoginType;
 import net.detalk.api.member.domain.MemberStatus;
 import net.detalk.api.member.domain.exception.MemberNotFoundException;
@@ -13,7 +14,7 @@ import net.detalk.api.member.controller.v1.response.GetMemberProfileResponse;
 import net.detalk.api.member.domain.Member;
 import net.detalk.api.member.domain.MemberExternal;
 import net.detalk.api.member.domain.MemberProfile;
-import net.detalk.api.repository.AttachmentFileRepository;
+import net.detalk.api.image.repository.AttachmentFileRepository;
 import net.detalk.api.repository.AuthRefreshTokenRepository;
 import net.detalk.api.member.repository.MemberExternalRepository;
 import net.detalk.api.member.repository.MemberProfileRepository;
