+++ src/main/java/net/detalk/api/support/s3/R2StorageClient.java
@@ -2,7 +2,7 @@ package net.detalk.api.support.s3;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import net.detalk.api.service.ImageService;
+import net.detalk.api.image.service.ImageService;
 import net.detalk.api.support.AppProperties;
 import org.springframework.stereotype.Component;
 import software.amazon.awssdk.services.s3.presigner.S3Presigner;
