+++ src/main/java/net/detalk/api/support/s3/S3Config.java
@@ -1,7 +1,7 @@
 package net.detalk.api.support.s3;
 
 import lombok.RequiredArgsConstructor;
-import net.detalk.api.service.ImageService;
+import net.detalk.api.image.service.ImageService;
 import net.detalk.api.support.AppProperties;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
