+++ src/test/java/net/detalk/api/service/FileServiceTest.java
@@ -5,7 +5,8 @@ import static org.assertj.core.api.Assertions.assertThat;
 import java.time.Instant;
 import java.time.LocalDateTime;
 import java.util.UUID;
-import net.detalk.api.domain.PreSignedData;
+import net.detalk.api.image.domain.PreSignedData;
+import net.detalk.api.image.service.FileService;
 import net.detalk.api.mock.FakeAttachmentFileRepository;
 import net.detalk.api.mock.FakeStorageClient;
 import net.detalk.api.mock.FakeTimeHolder;
